// interfaces/typeinterfaces.tsx
  export type Blog = {
    shortDescription: any;
    id: number;
    title: string;
    description: string;
    image: string;
    publishdate: number;
    isActive: string;
    metatitle: string;
    metaDescription: string;
    code: string;
  };